external/d4rllib/release/requirements.txt
# Requirements to run release tests from the local machine (including client tests)
click
# Copy anyscale pin to requirements_buildkite.txt and util.py
anyscale
slackclient
boto3
google-cloud-storage
jinja2
pybuildkite
PyGithub
pydantic < 1.10.0
pyyaml
typer[all]
toml
python-dotenv
PyGithub
expiringdict
requests
protobuf >= 3.15.3, != 3.19.5
pytz
retry
kubernetes
